+++ test/dummy/app/controllers/custom_namespace/resources_controller.rb
@@ -1,0 +1,5 @@
+module CustomNamespace
+  class ResourcesController < CustomNamespace::ApplicationController
+    include Layered::Resource::Controller
+  end
+end