+++ test/dummy/app/layered_resources/custom_namespace/post_resource.rb
@@ -1,0 +1,13 @@
+module CustomNamespace
+  class PostResource < Layered::Resource::Base
+    model ::Post
+
+    columns [
+      { attribute: :title, primary: true }
+    ]
+
+    fields [
+      { attribute: :title }
+    ]
+  end
+end